Migrate check-fea.py for fontbakery 0.12
